# idle: speed up test by 5x even while running 2x more iterations