Before this change, test iterations would last 500ms (* 10 iterations = 5s total).  With this change, test iterations happen in 50ms (*20 iterations = 1s).

Not super significant in and of itself, but it was a very simple change that cuts several seconds off per run (and we run each test 6 times in GA).

RELEASE NOTES: none